refactor: dynamic import 제거
