+++ src/app/layout.tsx
@@ -30,7 +30,7 @@ export default function RootLayout({
 				<head>
 					<meta
 						name="viewport"
-						content="width=device-width,initial-scale=1.0,maximum-scale=1.0,user-scalable=no"
+						content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no"
 					/>
 				</head>
 				<body
